pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.1",
    "discord-py>=2.6.4",
    "pynacl>=1.6.0",
    "python-dotenv>=1.1.1",
    "youtube-dl>=2021.12.17",
    "yt-dlp>=2025.10.14",
]
